GH-425 Avoid C++ compiler warning about conversion from string literal to 'char *.
In C string literals do not have the 'const' qualifier but in C++ they do.
Avoid C++ compiler warnings by changing our TRACE() funciton definition to use
'const char *' instead of plain 'char *'. This is a useful change for C too
because they string passed to the function can be expected to be read-only.